change admin_member model into user model
